Neural Net and Search Tree
The Neural Net is under Assign4 and the Search Tree is under Assign 2. They both have read me files attached. The neural net does not work in its current state. I need to fix the back propagation for updating the weight values. I plan on doing this in python later on.